+++ quantify_scheduler/backends/zhinst_backend.py
@@ -473,6 +473,7 @@ def _add_waveform_ids(timing_table: pd.DataFrame) -> pd.DataFrame:
         # waveform_id as it doesn't affect the waveform itself.
         waveform_op_id = re.sub(r"acq_index=\(.*\)", "acq_index=(*)", waveform_op_id)
         waveform_op_id = re.sub(r"acq_index=.*,", "acq_index=*,", waveform_op_id)
+        waveform_op_id = re.sub(r"acq_index=.*\)", "acq_index=*,", waveform_op_id)
 
         # samples should always be positive, the abs is here to catch a rare bug
         # where a very small negative number (e.g., -0.00000000000000013) is printed