+++ tests/scheduler/helpers/test_waveforms.py
@@ -13,6 +13,7 @@ from pytest_mock.plugin import MockerFixture
 
 from quantify_scheduler import Schedule
 from quantify_scheduler.helpers.schedule import get_pulse_uuid
+from quantify_scheduler.compilation import device_compile
 from quantify_scheduler.helpers.waveforms import (
     apply_mixer_skewness_corrections,
     area_pulse,
@@ -104,13 +105,14 @@ def test_get_waveform_by_pulseid(
 
 
 def test_get_waveform_by_pulseid_are_unique(
-    create_schedule_with_pulse_info: Callable,
+    load_example_transmon_config: Callable,
 ) -> None:
     # Arrange
     schedule = Schedule("my-schedule")
     schedule.add(X90("q0"))
     schedule.add(X90("q0"))
-    create_schedule_with_pulse_info(schedule)
+
+    schedule = device_compile(schedule, load_example_transmon_config())
 
     operation_repr = list(schedule.schedulables.values())[0]["operation_repr"]
     pulse_info_0 = schedule.operations[operation_repr]["pulse_info"][0]